refactor: ModalContent -> ContentArea 파일명 변경
